requirements.txt
Django==5.0.7
django-filter==24.3
django-import-export==4.1.1
djangorestframework==3.15.2
pre-commit==3.8.0
python-dotenv==1.0.1
virtualenv==20.26.3